refactor: extract duplicated authenticated_client fixture to conftest
Move the authenticated_client fixture from 3 identical definitions
inside describe blocks in test_push_spec.py to a shared
tests/core/conftest.py. All 30 tests pass unchanged.
